Curvature will be deprecated (read bellow)
I decided to drop curve resistance forces in the consumption algo. The current implementation is not exactly predictable with its parameter settings and previous iterations had their own flaws that lead to this state.

I changed the parameters in testing files (tram-exploration.ipynb and train-exploration.ipynb) so that it gets ignored for now. The results still look good despite this change. That leads me to believe that we can spare this feature.